Add horizonal and vertical DGModel kernels
This added kernels that operate in the reference space horizontal
directions and in the reference space vertical direction.

To minimize code duplication, the horizontal kernels are shared with the
every direction kernels.  In this implementation, the vertical kernels
are standalone in hopes that me might be able to optimize them further.